+++ bitframe-services/core/src/commonMain/kotlin/bitframe/service/config/ClientConfiguration.kt
@@ -1,8 +1,10 @@
 package bitframe.service.config
 
 import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.SupervisorJob
+import kotlin.jvm.JvmOverloads
 
-open class ClientConfiguration(
-    open val appId: String,
-    override val scope: CoroutineScope
-) : ServiceConfig()
+open class ClientConfiguration @JvmOverloads constructor(
+    override val appId: String,
+    override val scope: CoroutineScope = CoroutineScope(SupervisorJob())
+) : ServiceConfig(appId, scope)