+++ bitframe-services/core/src/commonMain/kotlin/bitframe/service/config/ServiceConfig.kt
@@ -6,12 +6,6 @@ import kotlin.jvm.JvmField
 import kotlin.jvm.JvmOverloads
 
 open class ServiceConfig @JvmOverloads constructor(
+    open val appId: String,
     open val scope: CoroutineScope = CoroutineScope(SupervisorJob())
-) {
-    companion object {
-        @JvmField
-        val DEFAULT = ServiceConfig(
-            scope = CoroutineScope(SupervisorJob())
-        )
-    }
-}
+)