+++ pi-monitor/pi-monitor-server/src/test/kotlin/unit/utils/Configuration.kt
@@ -3,13 +3,18 @@ package unit.utils
 import bitframe.ApplicationUnderTest
 import bitframe.server.InMemoryDaoProvider
 import bitframe.server.modules.authentication.AuthenticationServiceImpl
+import bitframe.service.config.ServiceConfig
 import pimonitor.PiMonitorServer
 import java.io.File
 
 @JvmField
 val DAO_PROVIDER_UNDER_TEST = InMemoryDaoProvider()
 
-val AUTH_SERVICE = AuthenticationServiceImpl(DAO_PROVIDER_UNDER_TEST)
+val SERVICE_CONFIG = ServiceConfig(
+    "server-app"
+)
+
+val AUTH_SERVICE = AuthenticationServiceImpl(DAO_PROVIDER_UNDER_TEST, SERVICE_CONFIG)
 
 val SERVER = ApplicationUnderTest(
     PiMonitorServer(